move traits to traits.rs
